+++ file/download.go
@@ -1,0 +1,20 @@
+package file
+
+import (
+	"io/ioutil"
+	"net/http"
+)
+
+func Download(url string) ([]byte, error) {
+	resp, err := http.Get(url)
+	if err != nil {
+		return nil, err
+	}
+
+	body, err := ioutil.ReadAll(resp.Body)
+	if err != nil {
+		return nil, err
+	}
+
+	return body, resp.Body.Close()
+}
